+++ changelog.md
@@ -36,6 +36,7 @@ Codion Change Log
 - Lambda module removed, worked fine for read operations, but didn't really make sense with writing and transactions.
 ### is.codion.swing.framework.ui
 - EditAttributePanel bug fixed, update no longer enabled if no entity is modified.
+- EntityTablePanel.Config.surrendersFocusOnKeystroke() added.
 ### is.codion.tools.generator.domain
 - DomainSource bug fixed, domain test imports fixed for api/impl.
 ### is.codion.tools.generator.panel